Update IrrigationController.ino
This includes changes allowing for a master valve or pump start control line.  The first output of the shift register is eliminated to make relay 1 the pump start relay.  Because of this, the values shifted out to the register need to be shifted one more bit over so that when valve 1 is called it starts on bit 2.

I also made a change to allow the pump start relay to be triggered on its own.  With this change, the pump is presented as another device to the controller.  When triggered, the display will show 
** Pump  only **
**     mode     **
When pump only mode is triggered, all sprinkler valves are shut off and only the pump is run.  A word of caution on this, if a separate irrigation pump is used with no automatic shutoff for the pump, it should never be run with no water flow happening.  I added this because in my setup I have an extra garden hose connected that I can use connected to my irrigation pump.